feature/1 코드 리팩토링
