+++ src/main/java/com/yscp/domain/strategy/CalculatorStrategy.java
@@ -1,0 +1,7 @@
+package com.yscp.domain.strategy;
+
+import com.yscp.domain.Number;
+
+public interface CalculatorStrategy {
+    Number calculate(String input);
+}
